refactor: simplify metric system - 48% backend reduction
- Merge metric.ts + metric-integration.ts into single unified router
- Create base.ts with 3 core functions (fetchData, renderEndpoints, getMetricTemplate)
- Split endpoints/ and metric-templates/ by integration (github, google-sheets, posthog, youtube)
- Merge youtube.ts + youtube-analytics.ts into single youtube integration
- Update dashboard.ts and youtube.ts routers to use new base functions
- Remove duplicate endpoint definitions across service files
- Delete old service files and fetch-integration-data.ts dispatcher
- Update all frontend metric creators with new API signatures
- Add proper TypeScript typing for transform functions

Backend: Reduced from ~3,787 lines to ~1,950 lines (48% reduction)

Files created:
- src/server/api/services/base.ts
- src/server/api/services/endpoints/{github,google-sheets,posthog,youtube}.ts
- src/server/api/services/metric-templates/{github,google-sheets,posthog,youtube}.ts

Files deleted:
- src/server/api/routers/metric-integration.ts
- src/server/api/services/{github,google-sheets,posthog,youtube,youtube-analytics,metric-templates,index}.ts
- src/server/api/utils/fetch-integration-data.ts
