+++ src/app/api-test/_components/integration-tester.tsx
@@ -74,30 +74,25 @@ export function IntegrationTester() {
   );
 
   // Fetch service configs for all supported services
-  const { data: githubService } =
-    api.metricIntegration.getServiceEndpoints.useQuery(
-      { integrationId: "github" },
-      { enabled: selectedIntegration?.integrationId === "github" },
-    );
-  const { data: googleSheetsService } =
-    api.metricIntegration.getServiceEndpoints.useQuery(
-      { integrationId: "google-sheet" },
-      { enabled: selectedIntegration?.integrationId === "google-sheet" },
-    );
-  const { data: posthogService } =
-    api.metricIntegration.getServiceEndpoints.useQuery(
-      { integrationId: "posthog" },
-      { enabled: selectedIntegration?.integrationId === "posthog" },
-    );
-  const { data: youtubeService } =
-    api.metricIntegration.getServiceEndpoints.useQuery(
-      { integrationId: "youtube" },
-      { enabled: selectedIntegration?.integrationId === "youtube" },
-    );
+  const { data: githubService } = api.metric.getServiceEndpoints.useQuery(
+    { integrationId: "github" },
+    { enabled: selectedIntegration?.integrationId === "github" },
+  );
+  const { data: googleSheetsService } = api.metric.getServiceEndpoints.useQuery(
+    { integrationId: "google-sheet" },
+    { enabled: selectedIntegration?.integrationId === "google-sheet" },
+  );
+  const { data: posthogService } = api.metric.getServiceEndpoints.useQuery(
+    { integrationId: "posthog" },
+    { enabled: selectedIntegration?.integrationId === "posthog" },
+  );
+  const { data: youtubeService } = api.metric.getServiceEndpoints.useQuery(
+    { integrationId: "youtube" },
+    { enabled: selectedIntegration?.integrationId === "youtube" },
+  );
 
   // Test endpoint mutation
-  const testMutation =
-    api.metricIntegration.testIntegrationEndpoint.useMutation();
+  const testMutation = api.metric.testIntegrationEndpoint.useMutation();
 
   // Get service config based on selected integration
   const serviceConfig = selectedIntegration?.integrationId
@@ -410,11 +405,8 @@ export function IntegrationTester() {
                             <Input
                               id={`${endpointKey}-${param}`}
                               placeholder={
-                                (
-                                  serviceConfig.exampleParams as
-                                    | Record<string, string>
-                                    | undefined
-                                )?.[param] ?? `Enter ${param}`
+                                serviceConfig.exampleParams?.[param] ??
+                                `Enter ${param}`
                               }
                               value={params[param] ?? ""}
                               onChange={(e) =>
