+++ src/app/metric/github/github-metric-creator.tsx
@@ -178,9 +178,10 @@ export function GitHubMetricCreator({
 
   // Fetch repositories dynamically
   const { data: repos, isLoading: loadingRepos } =
-    api.metricIntegration.fetchDynamicOptions.useQuery({
+    api.metric.fetchDynamicOptions.useQuery({
       connectionId,
-      endpoint: "github-repos",
+      templateId: repoTemplateId,
+      dropdownKey: "FULL_REPO",
     });
 
   // Create metric mutation
