+++ src/server/api/root.ts
@@ -2,7 +2,6 @@ import { dashboardRouter } from "@/server/api/routers/dashboard";
 import { feedbackRouter } from "@/server/api/routers/feedback";
 import { integrationRouter } from "@/server/api/routers/integration";
 import { metricRouter } from "@/server/api/routers/metric";
-import { metricIntegrationRouter } from "@/server/api/routers/metric-integration";
 import { organizationRouter } from "@/server/api/routers/organization";
 import { roleRouter } from "@/server/api/routers/role";
 import { taskRouter } from "@/server/api/routers/task";
@@ -21,7 +20,6 @@ export const appRouter = createTRPCRouter({
   team: teamRouter,
   role: roleRouter,
   metric: metricRouter,
-  metricIntegration: metricIntegrationRouter,
   integration: integrationRouter,
   dashboard: dashboardRouter,
   youtube: youtubeRouter,
